Migrated working code out of tests
