+++ src/Deserializer/WPP/TmfFileDecodingContextType.cs
@@ -4,22 +4,29 @@ using System.Text;
 namespace Nefarius.Utilities.ETW.Deserializer.WPP;
 
 /// <summary>
-///     A <see cref="TDH_CONTEXT_TYPE.TDH_CONTEXT_WPP_TMFFILE" />  wrapper for use with <see cref="DecodingContext" />.
+///     A <see cref="TDH_CONTEXT_TYPE.TDH_CONTEXT_WPP_TMFSEARCHPATH" />  wrapper for use with
+///     <see cref="DecodingContext" />.
 /// </summary>
 [SuppressMessage("ReSharper", "UnusedType.Global")]
-[Obsolete]
-public sealed class TmfFileDecodingContextType() : DecodingContextType(TDH_CONTEXT_TYPE.TDH_CONTEXT_WPP_TMFFILE)
+public sealed class TmfFileDecodingContextType()
+    : DecodingContextType(TDH_CONTEXT_TYPE.TDH_CONTEXT_WPP_TMFSEARCHPATH)
 {
     /// <summary>
-    ///     Gets decoding info from a single <c>.TMF</c> file.
+    ///     Gets decoding info from multiple paths containing <c>.TMF</c> files.
     /// </summary>
     /// <param name="path">
-    ///     Null-terminated Unicode string that contains the name of the .tmf file used for parsing the WPP log.
-    ///     Typically, the .tmf file name is picked up from the event GUID so you do not have to specify the file name.
+    ///     Null-terminated Unicode string that contains the path to the .tmf file. You do not have to
+    ///     specify this path if the search path contains the file. Only specify this context information if you also specify
+    ///     the TDH_CONTEXT_WPP_TMFFILE context type. If the file is not found, TDH searches the following locations in the
+    ///     given order:
+    ///     <ul>
+    ///         <li>The path specified in the TRACE_FORMAT_SEARCH_PATH environment variable</li>
+    ///         <li>The current folder</li>
+    ///     </ul>
     /// </param>
     public TmfFileDecodingContextType(string path) : this()
     {
-        ArgumentException.ThrowIfNullOrEmpty(path);
+        ArgumentNullException.ThrowIfNull(path);
         Buffer = new ReadOnlyMemory<byte>(Encoding.Unicode.GetBytes(Path.GetFullPath(path))
             .Concat("\0\0"u8.ToArray())
             .ToArray()
